pyproject.toml
[tool.poetry]
name = "gh-automation-funda"
version = "0.2.0"
description = "Funda.nl automation."
authors = ["Samuel Giffard <samuel@giffard.co>"]
license = "MIT"
readme = "README.md"
packages = [{include = "gh_automation_funda"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
pydantic = "^2.2.1"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
httpx = "^0.24.1"
beautifulsoup4 = "^4.12.2"

[tool.poetry.scripts]
auto = "gh_automation_funda.cli:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
black = "^23.3.0"
mypy = "^1.4.0"
pre-commit = "^3.3.3"
yoyo-migrations = "^8.2.0"
psycopg = {extras = ["c"], version = "^3.1.12"}
asyncpg-stubs = "^0.28.0"
psycopg-binary = "^3.1.10"
types-pygments = "^2.16.0.0"
types-beautifulsoup4 = "^4.12.0.7"

[tool.black]
line-length = 120

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
strict = "True"
plugins = "pydantic.mypy"
mypy_path = "stubs"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
